scripts/requirements.txt
autopep8==1.4.4
pycodestyle==2.5.0
PyYAML==5.1.1
tqdm==4.32.2
